Google sign in stop gap
The reason that Google doesn't load nicely when logging in is that the call
to `window.open` is too deeply nested in the callbacks. This fixes this issue
by introducing a second button. It requires that a user press the "click me first button", and then the authentication button. However, this __can__ be edited
so that it's done with one button / click using some clever jQuery techniques.